==> Step1-GenerateCorpus.ipynb <==
# Step 1: Generate corpus from all (JSON & PMC) files

import pandas as pd

metadata = pd.read_csv('../CORD19/metadata.csv')

# Select only those records having pdf_json or pmc_json or atleast abstract
metadata = metadata[(metadata['pdf_json_files'].notna() | metadata['pmc_json_files'].notna())]

metadata.tail()

metadata['date'] = pd.to_datetime(metadata['publish_time'].str.strip(), format='%Y-%m-%d') 

metadata = metadata[metadata['date'].dt.year == 2021]

metadata

len(metadata) - len(metadata[metadata['pdf_json_files'].notna() | metadata['pmc_json_files'].notna() | metadata['abstract'].notna()])

from tqdm import tqdm
from glob import glob
import json

pmc_count = 0
pdf_count = 0
results_list = []
for index, data in tqdm(metadata.iterrows(), total=len(metadata)):
    data_dict = dict()
    try:
        if pd.isna(data['pdf_json_files']):
            pmc_count += 1
            file_loc = '../CORD19/' + data['pmc_json_files']
            with open(file_loc) as json_file:
                target_json = json.load(json_file)
                data_dict['doc_id'] = target_json['paper_id']
                data_dict['title'] = target_json['metadata']['title']
                data_dict['cord_uid'] = data['cord_uid']
                abstract_section = str()
                data_dict['abstract'] = data['abstract']
                full_text_section = str()
                for element in target_json['body_text']:
                    full_text_section += element['text'] + ' '
                data_dict['full_text'] = full_text_section
        else:
            pdf_count += 1
            # PDF jsons file contains multiple docs separated by ";"
            file_loc = '../CORD19/' + data['pdf_json_files']
            file_loc = file_loc.split('; ')[0]
            with open(file_loc) as json_file:
                target_json = json.load(json_file)
                data_dict['doc_id'] = target_json['paper_id']
                data_dict['title'] = target_json['metadata']['title']
                data_dict['cord_uid'] = data['cord_uid']
                abstract_section = str()
                if target_json['abstract'] != '':
                    for element in target_json['abstract']:
                        abstract_section += element['text'] + ' '
                else:
                    abstract_section = data['abstract']
                data_dict['abstract'] = abstract_section
                full_text_section = str()
                for element in target_json['body_text']:
                    full_text_section += element['text'] + ' '
                data_dict['full_text'] = full_text_section
        results_list.append(data_dict)
    except Exception as msg:
        print('Error at: ' + data['pdf_json_files'])
        print(msg)

print('Total PMC records: {}'.format(pmc_count))
print('Total PDF records: {}'.format(pdf_count))

df_results = pd.DataFrame(results_list)
df_results.to_csv('total_corpus.csv', index=False)

df_results.tail()

df_results.describe()
