==> Step2-ExtractDiseaseNames.ipynb <==
# Step 2: Extract disease names from corpus

import pandas as pd
from tqdm import tqdm
from textsearch import TextSearch
import py_aho_corasick as ahoc
import numpy as np

df_results = pd.read_csv('total_corpus.csv')

df_results.iloc[int(len(df_results)-1)]

len(df_results)/2

corpus = ' '.join(df_results['full_text'].tolist())

corpus1 = ' '.join(df_results['full_text'][0:54763].tolist())

corpus2 = ' '.join(df_results['full_text'][54764:].tolist())

disease_data = pd.read_csv('../Data_Dictionary/DiseaseName.txt', delimiter = "\t", header=None, skipinitialspace=True)

disease_data.columns = ['DiseaseName']

disease_data['DiseaseName'] = disease_data['DiseaseName'].str.strip()

disease_list = disease_data['DiseaseName'].tolist()

disease_list[0:10]

corpus1[-100:].strip()

ts = TextSearch(case="ignore", returns="match", replace_foreign_chars=True)
ts.add(disease_list)

print(len(corpus1))
print(len(corpus2))

found_list1 = ts.findall(corpus1)

found_list1[-10:]

found_list2 = ts.findall(corpus2)

found_list2[-10:]

found_list = found_list1 + found_list2

found_list[-10:]

unique_found_list = list(dict.fromkeys(found_list))

mylist = [u'nowplaying', u'PBS', u'PBS', u'nowplaying', u'job', u'debate', u'thenandnow']
unique = list(dict.fromkeys(mylist))

unique_found_list[0:10]

with open('corona_related_disease_name_FINAL.txt', 'w') as filehandle:
    for listitem in unique_found_list:
        filehandle.write('%s\n' % listitem)

df_unique_found_list = pd.DataFrame(unique_found_list, columns=['disease'])

df_unique_found_list.tail()

df_unique_found_list.to_csv('disease_found_list.csv', index=False)

import os
print(os.cpu_count())

df_unique_found_list.disease.contains('baricitinib')
