==> Step3-ExtractDrugNames.ipynb <==
# Step 3: Extract drug names from corpus

import pandas as pd  # data processing, CSV file I/O (e.g. pd.read_csv)
import math
from textsearch import TextSearch

df_results = pd.read_csv('total_corpus.csv')

df_results.head()

#corpus = ' '.join(df_results['full_text'].tolist())

drug_data = pd.read_csv('../Data_Dictionary/Drugs.csv', encoding="utf-8", skipinitialspace=True)

drug_names = drug_data['Common name'].tolist()

drug_names[0]

corpus1 = ' '.join(df_results['full_text'][0:54763].tolist())
corpus2 = ' '.join(df_results['full_text'][54764:].tolist())

print(len(corpus1))
print(len(corpus2))

ts = TextSearch(case="ignore", returns="match", replace_foreign_chars=True)
ts.add(drug_names)

found_list1 = ts.findall(corpus1)

found_list2 = ts.findall(corpus2)

found_list = found_list1 + found_list2

unique_found_list = list(dict.fromkeys(found_list))

with open('corona_related_drug_name_FINAL.txt', 'w') as filehandle:
    for listitem in unique_found_list:
        filehandle.write('%s\n' % listitem)

unique_found_list[0:10]

df_unique_found_list = pd.DataFrame(unique_found_list, columns=['drug'])

df_unique_found_list.to_csv('drug_found_list.csv', index=False)

df_unique_found_list = pd.read_csv('drug_found_list.csv')

df_unique_found_list[df_unique_found_list.drug.str.contains('Baricitinib')]

drug_data[drug_data['Common name'].str.contains('Baricitinib', case=False)]

df_results[df_results.full_text.str.contains('paxlovid', case=False)]

df_results.iloc[7].full_text
