==> Step5-MapDrugNamesToArticle.ipynb <==
# Step 5: Map drug names to articles

import pandas as pd
from tqdm import tqdm

total_corpus = pd.read_csv('total_corpus.csv')

drug_df = pd.read_csv('drug_found_list.csv')

drug_df = drug_df[(drug_df['drug'] != 'Date') & (drug_df['drug'] != 'Water') & (drug_df['drug'] != 'Fig')]

drug_df.tail()

drug_list = drug_df['drug'].tolist()

drug_article_list = []

for drug in tqdm(drug_list):
    _list = ','.join(total_corpus[total_corpus['full_text'].str.contains(drug, case=False, regex=False, na=False)]['cord_uid'].tolist())
    if len(_list) > 0:
        drug_article_list.append([drug, _list])

drug_article_list[0:10]

drug_article_list_df = pd.DataFrame(drug_article_list, columns=['drug', 'article'])

drug_article_list_df.to_csv('drug_article_list.csv', index=False)
