==> Step6-DrugDiseaseCommonSentences.ipynb <==
# Step 6: Extract drug-disease common articles and sentences

import pandas as pd
import numpy as np
from tqdm import tqdm

disease_article_map = pd.read_csv('disease_article_list.csv')
drug_article_map = pd.read_csv('drug_article_list.csv')

drug_disease_list = []

for index, disease in tqdm(disease_article_map.iterrows(), total=len(disease_article_map)):
    disease_article = disease['article'].split(',')
    for i, drug in drug_article_map.iterrows():
        drug_article = drug['article'].split(',')
        common_article = list(set(disease_article) & set(drug_article))
        common_article_str = ','.join(common_article)
        if len(common_article_str) > 0:
            drug_disease_list.append([disease['disease'], drug['drug'], common_article_str])

## Extract common sentences where drug-disease pair is present

drug_disease_sentences = []

drug_disease_list[-10:]

drug_disease_list_df = pd.DataFrame(drug_disease_list, columns=['disease', 'drug', 'articles'])
drug_disease_list_df.to_csv('drug_disease_list.csv', index=False)

from nltk.tokenize import word_tokenize, sent_tokenize

total_corpus = pd.read_csv('total_corpus.csv')

total_corpus[total_corpus['cord_uid'] == 'rxz7w4r5'].full_text.values[0]

for row in tqdm(drug_disease_list):
    disease = row[0]
    drug = row[1]
    articles = row[2].split(',')
    for article in articles:
        content = str(total_corpus[total_corpus['cord_uid'] == article].full_text.values[0])
        chosen = []
        list1 = [str(drug).lower(), str(disease).lower()]
        for sentence in sent_tokenize(content):
            l = word_tokenize(sentence.lower())
            if len(set(l).intersection(list1)) > 1:
                chosen.append(sentence)
        if len(chosen) > 0:
            drug_disease_sentences.append([disease, drug, ''.join(chosen)])

drug_disease_sentences[-10:]
