==> Step6-DrugDiseaseCommonSentencesWithModin.ipynb <==
# Step 6: Extract drug-disease common articles and sentences

import os
os.environ["MODIN_ENGINE"] = "ray"  # Modin will use Ray

!pip install "modin[ray]"

import modin.pandas as pd

disease_article_map = pd.read_csv('disease_article_list.csv')
drug_article_map = pd.read_csv('drug_article_list.csv')

drug_disease_list = []

from tqdm import tqdm

drug_disease_list = pd.read_csv('drug_disease_list.csv')

drug_disease_list.head()

from nltk.tokenize import word_tokenize, sent_tokenize

total_corpus = pd.read_csv('total_corpus.csv')

import swifter

def process_disease_drug_rows(x):
    disease = x['disease']
    drug = x['drug']
    articles = x['articles'].split(',')
    for article in articles:
        content = str(total_corpus[total_corpus['cord_uid'] == article].full_text.values[0])
        chosen = []
        list1 = [str(drug).lower(), str(disease).lower()]
        for sentence in sent_tokenize(content):
            l = word_tokenize(sentence.lower())
            if len(set(l).intersection(list1)) > 1:
                chosen.append(sentence)
        if len(chosen) > 0:
            sentences = ''.join(chosen)
            return disease, drug, sentences
        else:
            return disease, drug, None

new_df = drug_disease_list.swifter.apply(process_disease_drug_rows, axis=1, result_type='expand')

new_df.describe()

new_df.columns = ['disease', 'drug', 'sentences']

new_df[new_df['sentences'].notna()]

drug_disease_sentences = new_df[new_df['sentences'].notna()]

drug_disease_sentences.describe()

drug_disease_sentences[-20:-10].values

drug_disease_sentences.to_csv('drug_disease_sentences.csv', index=False)

# for row in tqdm(drug_disease_list[0:10]):
#     disease = row[0]
#     drug = row[1]
#     articles = row[2].split(',')
#     for article in tqdm(articles):
#         content = str(total_corpus[total_corpus['cord_uid'] == article].full_text.values[0])
#         chosen = []
#         list1 = [str(drug).lower(), str(disease).lower()]
#         for sentence in sent_tokenize(content):
#             l = word_tokenize(sentence.lower())
#             if len(set(l).intersection(list1)) > 1:
#                 chosen.append(sentence)
#         if len(chosen) > 0:
#             drug_disease_sentences.append([disease, drug, ''.join(chosen)])
