==> Step7-CalculateDistanceBwDrugDisease.ipynb <==
# Step 7: Calculate distance between drug and disease

import os
os.environ["MODIN_ENGINE"] = "ray"  # Modin will use Ray
import modin.pandas as pd
import re

df_drug_disease_sentences = pd.read_csv('drug_disease_sentences.csv')

# Clean sentences before calculating distance:
def clean_sentence(text):
    return ''.join(e for e in text if (e.isalnum() or e.isspace()))

# Get minimum (absolute) distance between word1 and word1 in a sentence
def get_distance(w1, w2, sentence):
    words = sentence.split()
    if w1 in words and w2 in words:
          return abs(words.index(w2) - words.index(w1))
    return -1

def process_disease_drug_rows(data):
    word1 = clean_sentence(data['disease'].lower())
    word2 = clean_sentence(data['drug'].lower())
    sentence = clean_sentence(data['sentences'].lower())
    distance = get_distance(word1, word2, sentence)
    return distance

from tqdm import tqdm

df_drug_disease_sentences['distance'] = df_drug_disease_sentences.apply(process_disease_drug_rows, axis=1)

df_drug_disease_sentences.count()

df_drug_disease_sentences[df_drug_disease_sentences['distance'] == -1].values

df_drug_disease_sentences.to_csv('drug_disease_sentences_with_distance.csv', index=False)

## Appying on #1 & #2 dataset

dataframe = pd.read_csv('Dataset/dataset_1_junaed.csv')
dataframe2 = pd.read_csv('Dataset/dataset_2_hamada.csv')

dataframe2 = dataframe2[['Disease Name', 'Drug Name', 'Common Sentence', 'Manual Label']]

# Excluding -1 labels for binary classification
dataframe2 = dataframe2[dataframe2['Manual Label'] != -1]

dataframe = dataframe.append(dataframe2)

dataframe = dataframe[['Disease Name', 'Drug Name', 'Common Sentence', 'Manual Label']]

dataframe.describe()

Total 178 labels


dataframe.rename(columns={'Disease Name': 'disease', 'Drug Name': 'drug', 'Common Sentence': 'sentences'}, inplace=True)

dataframe = dataframe[dataframe.sentences.notna()]

dataframe['distance'] = dataframe.apply(process_disease_drug_rows, axis=1)

dataframe = dataframe[dataframe.distance != -1]

dataframe
