==> Step9-DiseaseDrugTextBlob.ipynb <==
# Step 9: Disease Drug Sentiment using TextBlob

# Applying textblob sentiment on sentences
import modin.pandas as pd
import csv
from textblob import TextBlob 
from sklearn.feature_extraction.text import TfidfVectorizer
from tqdm import tqdm

dataset = pd.read_csv('disease_drug_with_sentences_and_outlier_flag.csv')

dataset.tail()

def calculateAveragePolarity(row):
    drug = row['drug'].lower()
    disease = row['disease'].lower()
    sentences = row['sentences'].split('.')
    avg_polarity = 0
    total_polarity = 0
    for sentence in sentences:
        result = TextBlob(sentence.strip())
        total_polarity = total_polarity + result.polarity
    return total_polarity/len(sentences)
    

dataset['avg_polarity'] = dataset.apply(calculateAveragePolarity, axis=1)

dataset.describe()

dataset.to_csv('disease_drug_sentiment_textblob_and_outlier_flag.csv', index=False)
