==> Step10-UnsupervisedModelGenerateDataset.ipynb <==
from nltk.tokenize import word_tokenize
from nltk.corpus import stopwords
from nltk.stem.porter import PorterStemmer
import pandas as pd
from re import sub
import multiprocessing
import re
from gensim.models.phrases import Phrases, Phraser
from gensim.models import Word2Vec

def preprocess(text):
    text = str(text)

    #alpha numeric, punctuation remove kortesi
    text = sub(r"[^A-Za-z0-9^,!?.\/'+]", " ", text)
    text = sub(r"\+", " plus ", text)
    text = sub(r",", " ", text)
    text = sub(r"\.", " ", text)
    text = sub(r"!", " ! ", text)
    text = sub(r"\?", " ? ", text)
    text = sub(r"'", " ", text)
    text = sub(r":", " : ", text)
    text = sub(r"\s{2,}", " ", text)

    tokens = word_tokenize(text)
    stop_words = set(stopwords.words('english'))
    tokens = [w for w in tokens if w == 'no' or w == 'not' or not w in stop_words]
    porter = PorterStemmer()
    stems = []
    for t in tokens:
        stems.append(porter.stem(t))

    clean_words = []
    rx = re.compile(r'\D*\d')
    for s in stems:
        if rx.match(s):
            continue 
        clean_words.append(s)
    return clean_words

dataset = pd.read_csv("disease_drug_sentiment_textblob_and_outlier_flag.csv")

disease_list = dataset['disease'].tolist()
drug_list = dataset['drug'].tolist()
article_list = dataset['sentences'].tolist()

processed_article_list = []

for article in article_list:
    processed_article = preprocess(article)
    processed_article_list.append(processed_article)

sent = processed_article_list #just to keep it as source
phrases = Phrases(sent, min_count=1, progress_per=50000)
bigram = Phraser(phrases)
sentences = bigram[sent]

w2v_model = Word2Vec(min_count=5, window=4, vector_size=300, sample=1e-5, alpha=0.03, min_alpha=0.0007, negative=20, workers=multiprocessing.cpu_count())

w2v_model.build_vocab(sentences, progress_per=50000)

w2v_model.train(sentences, total_examples=w2v_model.corpus_count, epochs=30, report_delay=1)

w2v_model.init_sims(replace=True)

w2v_model.save("word2vec.model")

new_article_list = []
for article in processed_article_list:
    new_article = ' '.join(bigram[article])
    new_article_list.append(new_article)

zippedList =  list(zip(disease_list, drug_list, new_article_list))
clean_data = pd.DataFrame(zippedList, columns = ['disease' , 'drug', 'title']) 
clean_data.to_csv('cleaned_dataset.csv', index=False)

clean_data.head()

ab_dataset = pd.read_csv('Dataset/dataset_3_abdullah.csv')

ab_dataset.rename(columns={'chosen_sentence': 'sentences'}, inplace=True)

article_list_ab = ab_dataset['sentences'].tolist()

processed_article_list_ab = []

for article in article_list_ab:
    processed_article = preprocess(article)
    processed_article_list_ab.append(processed_article)

new_article_list_ab = []
for article in processed_article_list_ab:
    new_article = ' '.join(bigram[article])
    new_article_list_ab.append(new_article)

ab_dataset['title'] = new_article_list_ab

ab_dataset.to_csv('Dataset/dataset_3_abdullah.csv', index=False)

## Get title values for combined dataset

combined_dataset = pd.read_csv('Dataset/Combined_Dataset_2022-18-02.csv')

article_list_combined = combined_dataset['sentences'].tolist()

processed_article_list_combined = []

for article in article_list_combined:
    processed_article = preprocess(article)
    processed_article_list_combined.append(processed_article)

new_article_list_combined = []
for article in processed_article_list_combined:
    new_article = ' '.join(bigram[article])
    new_article_list_combined.append(new_article)

combined_dataset['title'] = new_article_list_combined

combined_dataset.to_csv('Dataset/Combined_Dataset_2022-18-02.csv', index=False)
