==> Step15-MakingGeneDataFrame.ipynb <==
import pandas as pd
import numpy as np

lrna_dataset = pd.read_csv('Final_lncRNA.csv')
mirna_dataset = pd.read_csv('Final_miRNA.csv')

lrna_dataset['type'] = 'lncRNA'
mirna_dataset['type'] = 'miRNA'
lrna_dataset.rename(columns={lrna_dataset.columns[0]:'approved_symbol'}, inplace=True)
mirna_dataset.rename(columns={mirna_dataset.columns[0]:'approved_symbol'}, inplace=True)

mirna_dataset.columns

lrna_dataset.columns

full_gene_df = lrna_dataset.append(mirna_dataset, ignore_index=True)

full_gene_df.drop_duplicates(subset=['approved_symbol'], keep='first')

serial_list = range(1, len(full_gene_df)+1)
gene_id_list = ["G"+str(s) for s in serial_list]

full_gene_df['geneId'] = gene_id_list

full_gene_df.to_csv("full_gene_df.csv", index=False)
