==> Step16-AbstractPaperForGene.ipynb <==
import pandas as pd
import numpy as np

metadata = pd.read_csv('../CORD19/metadata.csv')

metadata = metadata[['title', 'doi', 'abstract']]

serial_list = range(1, len(metadata)+1)
metadata['metaId'] = serial_list

metadata.head()

metadata.to_csv('abstract_paper_for_gene.csv', index=False)
