==> Step17-GeneAbstractMap.ipynb <==
import pandas as pd
from tqdm import tqdm
import numpy as np

abstract_paper = pd.read_csv('abstract_paper_for_gene.csv')

full_gene_df = pd.read_csv('full_gene_df.csv')

abstract_paper.dropna(inplace=True)

abstract_paper.describe()

gene_list = full_gene_df['approved_symbol'].tolist()

abstract_paper['Keyword1'] = abstract_paper['abstract'].apply(lambda x: ', '.join(set(x.split()).intersection(set(gene_list))))

abstract_paper[abstract_paper['Keyword1'] != '']

disease_list = pd.read_csv('disease_found_list.csv')

disease_list.head()

disease_set = disease_list.disease.tolist()

abstract_paper['Keyword2'] = abstract_paper['abstract'].apply(lambda x: ', '.join(set(x.split()).intersection(set(disease_set))))

gene_disease_abstract = abstract_paper[(abstract_paper['Keyword2'] != '') & (abstract_paper['Keyword1'] != '')]

gene_disease_abstract.tail()

gene_disease_abstract_map = gene_disease_abstract.assign(Keyword2=gene_disease_abstract.Keyword2.str.split(',')).explode(['Keyword2'])

gene_disease_abstract_map = gene_disease_abstract_map.assign(Keyword1=gene_disease_abstract.Keyword1.str.split(',')).explode(['Keyword1'])

gene_disease_abstract_map.reset_index(inplace=True)

gene_disease_abstract_map = gene_disease_abstract_map.rename(columns={'Keyword1': 'gene', 'Keyword2': 'disease'})

gene_disease_abstract_map = gene_disease_abstract_map.drop(columns=['index'], axis=1)

gene_disease_abstract_map.dropna(subset=['gene', 'disease'])

gene_disease_abstract_map.to_csv('gene_disease_abstract_map.csv', index=False)
