==> Step19-CosineSimilarity_Word2Vec_GeneDisease.ipynb <==
from gensim.models import Word2Vec
import pandas as pd

word_vectors = Word2Vec.load("word2vec_for_gene.model").wv

gene_disease_abstract_map = pd.read_csv('gene_disease_abstract_map_with_stems.csv')

def compute_cosine(row):
    disease = row.disease
    gene = row.gene
    try:
        cosine_similarity = word_vectors.similarity(disease, gene)
    except:
        try:
            cosine_similarity = word_vectors.similarity(disease, gene.lower())
        except:
            cosine_similarity = 'UNKNOWN'
    return cosine_similarity

gene_disease_abstract_map['cosine'] = gene_disease_abstract_map.apply(compute_cosine, axis=1)

gene_disease_abstract_map[gene_disease_abstract_map.cosine == 'UNKNOWN']

word_vectors.most_similar('COVID-19')

gene_disease_abstract_map.to_csv('gene_disease_abstract_map_with_cosine.csv', index=False)

gene_disease_abstract_map.describe(include='all')

dg_ext = pd.read_csv('../August/Dictionary/all_gene_disease_associations.tsv', sep='\t', header=0)

dg_ext.drop_duplicates()

dg_cosine = gene_disease_abstract_map[['disease', 'gene', 'cosine', 'title', 'abstract', 'doi']]

dg_ext['disease'] = dg_ext.diseaseName.str.lower()
dg_ext['gene'] = dg_ext.geneSymbol

merged_df = dg_ext.merge(dg_cosine, on=['disease', 'gene'])

merged_df

dg_cosine['found_flag'] = 0
merged_df['found_flag'] = 1

cosine_similarity_known_list = merged_df['cosine']

cosine_similarity_known_list = [float(x) for x in cosine_similarity_known_list] 

avg_cosine_known = sum(cosine_similarity_known_list) / len(cosine_similarity_known_list)
max_cosine_known = max(cosine_similarity_known_list)
min_cosine_known = min(cosine_similarity_known_list)

print("min known cosine:")
print(min_cosine_known)
print("avg known cosine:")
print(avg_cosine_known)
print("max known cosine:")
print(max_cosine_known)

main_df = dg_cosine.append(merged_df[['disease', 'gene', 'cosine', 'found_flag', 'title', 'abstract', 'doi']], ignore_index=True)

main_df = main_df.sort_values('found_flag', ascending=False).drop_duplicates(subset=['disease', 'gene'])

main_df

def compute_association(row):
    association_type = 'ERROR'
    found_flag = row['found_flag']
    if found_flag == 1:
        association_type = 'Verified'
    else:
        cosine = row['cosine']
        if cosine == 'UNKNOWN':
            association_type = 'Low'
        else:
            cosine = float(cosine)
            distance_from_max_known = abs(max_cosine_known - cosine)
            distance_from_avg_known = abs(avg_cosine_known - cosine)
            distance_from_min_known = abs(min_cosine_known - cosine)
            min_distance = min(distance_from_max_known, distance_from_avg_known, distance_from_min_known)
            if min_distance == distance_from_max_known:
                association_type = 'High'
            elif min_distance == distance_from_avg_known:
                association_type = 'Medium'
            elif min_distance == distance_from_min_known:
                association_type = 'Low'
    return association_type

main_df['association'] = main_df.apply(compute_association, axis=1)

main_df.to_csv("gene_disease_association_with_confidence.csv", index=False)
