==> Step20-DrugPDBAssociation.ipynb <==
import pandas as pd
import numpy as np

from nltk.tokenize import word_tokenize, sent_tokenize

from textsearch import TextSearch

metadata = pd.read_csv('../CORD19/metadata.csv')

metadata = metadata[['title', 'doi', 'abstract']]

metadata['metaId'] = range(1, len(metadata) + 1)

pdb_ids = pd.read_csv('../August/Dictionary/Final_PDB_2011_11_03.txt', header=None)

pdb_ids.columns = ['pdb']

metadata = metadata[metadata.abstract.notna()]

len(metadata)/2

abstract_list_1 = ' '.join(metadata[0:284512]['abstract'].tolist())
abstract_list_2 = ' '.join(metadata[284513:]['abstract'].tolist())

ts = TextSearch(case='ignore', returns='match')
ts.add(pdb_ids.pdb.tolist())
found_pdbs_1 = ts.findall(abstract_list_1)

found_pdbs_2 = ts.findall(abstract_list_2)

found_pdbs_2.extend(found_pdbs_1)

unique_pdbs = list(dict.fromkeys(found_pdbs_2))

len(unique_pdbs)

## Starting drug search

# Get drugs from list
drugs_list = pd.read_csv('../Data_Dictionary/Drugs.csv')

drugs_list

ts = TextSearch(case='ignore', returns='match')
ts.add(drugs_list['Common name'].tolist())
found_drugs = ts.findall(abstract_list_1)

found_drugs_1 = ts.findall(abstract_list_2)

found_drugs_1.extend(found_drugs)

unique_drugs = list(dict.fromkeys(found_drugs_1))

unique_drugs

unique_drugs_df = pd.DataFrame(unique_drugs, columns=['drug'])

unique_pdbs_df = pd.DataFrame(unique_pdbs, columns=['pdb'])

unique_pdbs_df.to_csv('Found_Lists/pdb_found_list.csv', index=False)

def find_pdb_in_metadata(row):
    pdbs = unique_pdbs_df.pdb.tolist()

unique_pdbs_df

unique_drugs_df[unique_drugs_df.drug.str.contains('hydrocortisone', case=False)]

metadata['Keyword1'] = metadata['abstract'].apply(lambda x: ', '.join(set(x.split()).intersection(set(unique_drugs_df.drug.tolist()))))

metadata['Keyword2'] = metadata['abstract'].apply(lambda x: ', '.join(set(x.split()).intersection(set(unique_pdbs_df.pdb.tolist()))))

found_pdb_drug = metadata[(metadata.Keyword1 != '') & (metadata.Keyword2 != '')]

found_pdb_drug

drug_pdb_abstract_map = found_pdb_drug.assign(Keyword1=found_pdb_drug.Keyword1.str.split(',')).explode(['Keyword1'])

drug_pdb_abstract_map = drug_pdb_abstract_map.assign(Keyword2=found_pdb_drug.Keyword2.str.split(',')).explode(['Keyword2'])

drug_pdb_abstract_map.reset_index(drop=True)

drug_pdb_abstract_map.to_csv('Found_Lists/drug_pdb_abstract_map.csv', index=False)

unique_drugs
